Add Routes and link pages through routes
