+++ src/routes/LogbookPage.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+import LogEntryList from '../components/LogEntryList/LogEntryList';
+import Button from '../components/Button/Button';
+
+export default function LogbookPage({ logEntries }) {
+  return (
+    <>
+      <LogEntryList logEntries={logEntries} />
+      <Button onClick={handleClick} variant="navigate">
+        create a new log entry
+      </Button>
+    </>
+  );
+
+  function handleClick() {}
+}
